Create pong.js
Tried to make the puck start again at the middle of the screen but couldn't figure out how to make the function wait before starting over again to let the user know that they scored a point. Went for the simple solution and it makes the users have to press play again before starting another match.